Fix swapping out Database in tests
Previously overwrote the object itself for a quick swap of DatabaseInitialization, but that only works over API 24.
So now we will use loadKoinModules instead, which resolves the issue on ani API level
